Changed HelperData filed to long (from int)

+++ src/org/supremica/automata/algorithms/AutomataSynchronizer.java
@@ -132,7 +132,7 @@ public class AutomataSynchronizer
 	}
 
 	// -- MF -- Added to allow users easy access to the number of synch'ed states
-	public int getNumberOfStates()
+	public long getNumberOfStates()
 	{
 		return synchHelper.getNumberOfAddedStates();
 	}
